requirements.txt
PyJWT
certifi==2018.8.24
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-dotenv==0.9.1
requests==2.19.1
urllib3==1.23
Werkzeug==0.14.1
